+++ alpine.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+docker run --rm -it \
+	-v ./src:/opt/installer \
+	-w /opt/installer \
+	alpine sh